% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{add_attribute_set}
\alias{add_attribute_set}
\title{Title}
\usage{
add_attribute_set(attribute, iscell)
}
\arguments{
\item{attribute}{julia object of attribute (the result of create_attribute or quick_training)}

\item{iscell}{pre-existing attribute set (julia object)}
}
\value{
merged attribute set
}
\description{
adding attribute object into attribute set
}
\examples{

}
