% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{add_signature_set}
\alias{add_signature_set}
\title{Title}
\usage{
add_signature_set(signature, mscell)
}
\arguments{
\item{signature}{julia object of signature (the result of create_signature)}

\item{mscell}{per-existing signature set}
}
\value{
merged signature set
}
\description{
adding signature object into signature set
}
\examples{

}
