% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{convert_seurat}
\alias{convert_seurat}
\title{convert_seurat}
\usage{
convert_seurat(seurat_object, slot_option = "counts")
}
\arguments{
\item{slot_option}{slot parameter in GetAssayData function in Seurat package (should be "counts", otherwise, prediction wouldn't show good result)}

\item{seurtat_object}{R variable which is a seurat object}
}
\value{
a matrix class composed of gene(row)-by-cell(column)
}
\description{
covert seurat object into matrix format (same result format as importdata function)
}
\examples{
data <- convert_seurat(seurat_obj)

}
