% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{create_attribute}
\alias{create_attribute}
\title{Title}
\usage{
create_attribute(Name, Tissue, w2v, Expr, Genename, Cellname,
  cell_ontology, Sdatatype, w2v_index = FALSE, Xdatatype = "scRNA-seq",
  Ngene = 64, Stopwords = stopwords, Fdrthr = 0.1, Gamma = 1,
  Lambda = 1, Xpeaks = FALSE, Sep = "_", Mincellcount = 10,
  Mincellfrac = 0.05, Simthres = -Inf, R_output = FALSE)
}
\arguments{
\item{Name}{user-defined name for the output}

\item{Tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea}

\item{w2v}{word2vector object
output of make_w2v
pre-existing side-information object (julia object) (ex: julia_obj$side_information) 
output of generate_sideinformation}

\item{Expr}{expression/peak matrix (first index from the result of make_input: d <- make_input(input_var), d[[1]])}

\item{Genename}{gene/peak_location of Expr (second index from the result of make_input: d<- make_input(input_var), d[[2]])}

\item{Cellname}{cellname of Expr (third index from the result of make_input: d<- make_input(input_var), d[[3]])}

\item{cell_ontology}{julia object of cell ontology (the result of make_cellontology)}

\item{Sdatatype}{type of side-information (cell_ontology, ATAC-seq, or user-defined term)}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object for side-information exists (default: FALSE)}

\item{Xdatatype}{type of Expr (default: scRNA-seq; or scATAC-seq)}

\item{Ngene}{a number of vector size after JL transformation (default: 64)}

\item{Stopwords}{julia object of stopwords (the result of make_stopwords) (default: pre-existing variable: stopwords)}

\item{Fdrthr}{fdr threshold for calculating similarity score during training (default: 0.1)}

\item{Gamma}{paramter for ZSL (default: 1)}

\item{Lambda}{parameter for ZSL (default: 1)}

\item{Xpeaks}{TRUE if data type of x (or Expr) is ATAC-seq (default: FALSE)}

\item{Sep}{delimiter for peak_location (default: "_")
ex: chr1_3084739_3085712 (make sure delimiter for peak_location of data must be "_")}

\item{Mincellcount}{minimum cell count for a given cell type for training (default: 10)}

\item{Mincellfrac}{minimum fraction of cell type for training (default: 0.05)}

\item{Simthres}{initial similarity threshold (default: -Inf)}

\item{R_output}{TRUE ifthe  result of the function should be R object. FALSE if the result of the function should be julia object (default: FALSE)}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
make attribute object for ZSL (trained data)
}
\examples{

}
