% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{create_meta_index_av}
\alias{create_meta_index_av}
\title{create_meta_index_av}
\usage{
create_meta_index_av(iscell, mscell, gn_overlap_thr = 0.5, ncore = 1)
}
\arguments{
\item{iscell}{attribute set}

\item{mscell}{signature set}

\item{gn_overlap_thr}{minimum gene overlap ratio of attribute in attribute set between query data}

\item{ncore}{number of cores for multi-threading}
}
\value{
return list composed of (model_list, sname, zsl_tissue, meta_column_name)
\itemize{
\item model_list: logistic regression model (composed of predicting cell and corresponding coefficients)
\item sname: cell types that are used in each coefficient (T cell, B cell, fibroblast ....)
\item zsl_tissue: tissue types that are used in logistic regression (Lung, Heart ...)
\item meta_column_name: pseudo name of variables in logistic regression (Heart_cell1, Heart_cell2...)
}
}
\description{
creating logistic regression model (data of iscell and mscell must be independent)
}
\examples{
meta_model <- create_meta_index_av(att_set, sig_set)

}
