% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{create_signature}
\alias{create_signature}
\title{create_signature}
\usage{
create_signature(
  Name,
  Tissue,
  Expr,
  Rowname,
  Cellname,
  side_info,
  w2v_index = TRUE,
  snames = c(),
  stypenames = c(),
  Xdatatype = "scRNA-seq",
  sdatatype = "w2v",
  R_output = FALSE,
  celltype_cnt_thr = TRUE,
  mincellfrac = 0,
  mincellcount = 10,
  side_limit = FALSE
)
}
\arguments{
\item{Name}{user-defined name for the output}

\item{Tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
\itemize{
\item Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea
}}

\item{Expr}{expression/peak matrix (d <- make_input(input_var), d$norm_matrix)}

\item{Rowname}{gene/peak_location of Expr (d<- make_input(input_var), d$rowname)}

\item{Cellname}{cellname of Expr (d<- make_input(input_var), d$cellname)}

\item{side_info}{matrix or w2v object depends on w2v_index
\itemize{
\item (gene or peak) matrix of side information (if w2v_index = FALSE)
\item w2v_side: result of make_w2v_sideinfo (if w2v_index = TRUE)
}}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object of side-information exists}

\item{snames}{gene/peak_location of matrix for side information (if w2v_index = FALSE)}

\item{stypenames}{cellname of matrix for side information (if w2v_index = FALSE)}

\item{Xdatatype}{type of Expr
\itemize{
\item scRNA-seq
\item scATAC-seq
}}

\item{sdatatype}{type of side-information
\itemize{
\item w2v (cell ontology)
\item scRNA-seq
\item scATAC-seq
}}

\item{R_output}{TRUE if the result of the function should be R object. FALSE if the result of the function should be julia object}

\item{celltype_cnt_thr}{if the number for cell type exceeds, warning message will be sent and return FALSE. To ignore it, change the parameter into FALSE}

\item{Mincellcount}{minimum cell count for a given cell type (works when side_limit: TRUE)}

\item{Mincellfrac}{minimum fraction of cell type (works when side_limit: TRUE)}

\item{side_liimit}{TRUE if user wants to limit their side information as cell types in input data. Otherwise, it will use all the cell types in side information (= FALSE)}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
make signature object for ZSL (input for prediction)
}
\examples{
1) Side information: Cell Ontology
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
w2v_side <- make_w2v_sideinfo(w2v, annot, stopwords)
query_sig <- create_signature("user", "Heart", d$norm_matirx, d$rowname, d$cellname, w2v_side)

2) Side information: scRNA-seq (or scATAC-seq, etc...)
d <- make_input(matrix1, train = TRUE)
sideinfo <- make_input(matrix2, train = TRUE)
query_sig <- create_signature("user", "Heart", d$norm_matirx, d$rowname, d$cellname, sideinfo$norm_matrix, w2v_index = FALSE, snames = sideinfo$rowname, stypenames = sideinfo$cellname, sdatatype = "scRNA-seq")

}
