% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{decorate_all_confusion}
\alias{decorate_all_confusion}
\title{decorate_all_confusion}
\usage{
decorate_all_confusion(all_confusion)
}
\arguments{
\item{all_confusion}{all_confusion matrix from "transfer_map_consensus" function in Julia (which has rowname and colname)}
}
\value{
summarized format of all_confusion matrix
}
\description{
make a summarized format of all_confusion matrix (internal function of predict_cell)
}
\examples{
NA

}
