% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{estimate_celltype}
\alias{estimate_celltype}
\title{estimate_celltype}
\usage{
estimate_celltype(conf)
}
\arguments{
\item{conf}{confusion matrix from "predict_cell"}
}
\value{
suggests matched cell type for corresponding query cell type with q-value, respectively
}
\description{
prediction of cell type by a given confusion matrix from "predict_cell". Assume binomial distribution: n: total cell count for each query celltype / p: 1/celltypes from cell ontology / X (for z score): cell count for each celltype from cell ontology. Only consider max cell count (draw -> mixed)
}
\examples{
confusion <- predict_cell(att_set, query)
estimated_cell <- estimate_celltype(confusion)

}
