% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{generate_sideinformation}
\alias{generate_sideinformation}
\title{Title}
\usage{
generate_sideinformation(Side_mat, Cellnames, Sdim = 200)
}
\arguments{
\item{Side_mat}{matrix class composed of gene/peak_location(row)-by-cell(column) (same as result of importdata or convert_seurat)}

\item{Sdim}{the size of side-information for each cell type (default: 200)}

\item{cellnames}{cell name of Side_mat (column name)}
}
\value{
a side-information which could be directly used in create_siganture or create_attribute with w2v_index = FALSE (julia_object)
}
\description{
Generation of side-information from expression or peak matrix
}
\examples{
d <- make_input(input_var)
s <- generate_sideinformation(d[[1]], d[[3]])
}
