% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{import_sparse}
\alias{import_sparse}
\title{import_sparse}
\usage{
import_sparse(address, suffix_file, row_opt = 1)
}
\arguments{
\item{address}{scdata address list (contains only the prefix of the file: Heart, Lung ...)}

\item{suffix_file}{suffix for rowname_file, cellname_file, and sparse matrix
\itemize{
\item format of the suffix_file (should be same order as shown below)
\item _cells.tsv
\item _row.tsv
\item _count.mtx
}}

\item{row_opt}{indicates the column number that sholud be used for the gene(peak) name}
}
\value{
list of matrix (names: each tissue or filename)
}
\description{
"importdata" for TM (or multiple "sparse" matrix files)
}
\examples{
tm_sparse <- import_sparse("address_file", "suffix_file")

}
