% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{import_tm_facs}
\alias{import_tm_facs}
\title{import_tm_facs}
\usage{
import_tm_facs(address)
}
\arguments{
\item{address}{scdata address list (format: "raw")\cr
- filename: tissue_1_cnt, tissue_2_cnt ...}
}
\value{
list of matrix (names: each tissue or filename)
}
\description{
"importdata" for TM (or multiple "raw" matrix files)
}
\examples{
tm_facs <- import_tm_facs("address_file")

}
