% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{importdata}
\alias{importdata}
\title{importdata}
\usage{
importdata(file, opt = "raw", datatype = "RNAseq",
  colname_file = FALSE, rowname_file = FALSE, row_opt = 1)
}
\arguments{
\item{file}{file of the single-cell data (ex: scRNA-seq, scATAC-seq)}

\item{opt}{data format of the single-cell data
\itemize{
\item raw: tab-delimited text file
\item SingleCellExperiment: SingleCellExperiment rds file
\item sparse: MatrixMarket format (written by writeMM)
}}

\item{datatype}{"RNAseq" or "atac"}

\item{colname_file}{colname_file for sparse matrix (usually cell name), when opt is "sparse". Otherwise, it is FALSE}

\item{rowname_file}{rowname_file for sparse matrix (gene name: official gene symbol or peak location), when opt is "sparse". Otherwise, it is FALSE}

\item{row_opt}{indicates the column number that sholud be used for the gene(peak) name}
}
\value{
a matrix class composed of gene/peak_location(row)-by-cell(column)
}
\description{
importdata will import sc-data (scRNA-seq or scATAC-seq) and make gene(peak)-by-cell matrix
}
\examples{
data <- importdata("filename")
- scRNA-seq with raw matrix file
data <- importdata("sparse matrix file name", opt = "sparse", colname_file = "cell file", rowname_file = "row_file")
- for sparse matrix file

}
