% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{julia_environment}
\alias{julia_environment}
\title{Title}
\usage{
julia_environment(julia_obj_index = FALSE)
}
\arguments{
\item{julia_obj_index}{it should be TRUE if importing pre-existing julia object (default FALSE)}
}
\value{
return the julia_object if julia_obj_index is TRUE. Otherwise, there is no output
}
\description{
setup the environment for calling Julia packages and Julia code (CoreFunction.jl) (importing pre-existing julia object is optional)
}
\examples{

}
