% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{leave_one_val}
\alias{leave_one_val}
\title{Title}
\usage{
leave_one_val(att_set, sig_set)
}
\arguments{
\item{att_set}{attribute set}

\item{sig_set}{signature set}
}
\value{
res_list, ratio_list, val_list, tissue_list, f1_list, unassign_list
res[[1]]: res_list (confusion matrix)
res[[2]]: ratio_list (confusion_ratio_matrix)
res[[3]]: val_list (validation matrix)
res[[4]]: tissue_list
res[[5]]: f1_list (f1 score)
res[[6]]: unassign_list
each result is composed of sample in TM (ex: res[[1]][1]: confusion matrix of Aorta tissue)
}
\description{
leave one tissue validation of TM (the parameter for side-information of att_set and sig_set must be same)
}
\examples{

}
