% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{logistic_prediction_cell}
\alias{logistic_prediction_cell}
\title{Title}
\usage{
logistic_prediction_cell(iscell, meta_model_list, query,
  gn_overlap_thr = 0.5, logistic_thr = 0.5, tissue_opt = FALSE,
  celltype_specific = FALSE)
}
\arguments{
\item{iscell}{attribute set}

\item{meta_model_list}{result from create_meta_index}

\item{query}{signature object of query}

\item{gn_overlap_thr}{minimum gene overlap ratio of attribute in attribute set between query data (default: 0.5)}

\item{logistic_thr}{threshold for each cell whether predicted score from logistic regression model is confident (default: 0.5)}

\item{tissue_opt}{TRUE if user wants to find tissue-origin (default: FALSE)}

\item{celltype_specific}{TRUE if user wants to calculate the weight fo each tissue with only matched cell type. Otherwise, weight for each tissue will be calculated with whole celltypes (default: FALSE; tissue_opt should be TRUE to use this parameter)}
}
\value{
return 2 confusion matrix 
res[[1]]: a confusion matrix of prediction result
row: query cell type or cluster name
column: cell types in side-information (from query)
first column: unassign count
res[[2]]: a confusion matrix of prediction result in each cell
row: query cell
}
\description{
predict cell type with logistic regression model
}
\examples{

}
