% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{make_attribute_set}
\alias{make_attribute_set}
\title{make_attribute_set}
\usage{
make_attribute_set(attribute)
}
\arguments{
\item{attribute}{julia object of attribute (the result of create_attribute or quick_training)}
}
\value{
initiallized attribute set
}
\description{
making attribute set
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
att <- create_attribute("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)
att_set <- make_attribute_set(att)

}
