% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{make_input}
\alias{make_input}
\title{make_input}
\usage{
make_input(input_var, min_thr = 0.1, max_thr = 0.9, pseudo_cnt = 1,
  train = FALSE, gene_usage = FALSE, norm = "log")
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column)\cr
- same as result of importdata or convert_seurat}

\item{min_thr}{minimum threshold of expressing cell ratio for a given gene/peak_location}

\item{max_thr}{maximum threshold of expressing cell ratio for a given gene/peak_location}

\item{pseudo_cnt}{pseudo count for count matrix for log-transformation}

\item{train}{TRUE if it is data for training
\itemize{
\item TRUE: adjust min_thr and max_thr. In addition, make the gene/peak_location count for hadamard matrix
\item FALSE: no gene/peak_location filtering
}}

\item{gene_usage}{option for gene/peak_location filtering to make hadamard matrix
\itemize{
\item FALSE: gene/peak_location filtering by discarding low expressing cell ratio genes/peak_locations
\item certain number (ex: 2048): the number of gene to use for training (must be suitable number to generate hadamard matrix)
}}

\item{norm}{normalization options for matrix
\itemize{
\item "log": log-normalization (with pseudo count)
\item "raw": just uses count matrix
\item "total": divide by total count of each cell, respectively
\item "z": z-normalization
\item "total-log": "total" -> "log" normalization (with pseudo count)
}}
}
\value{
return a list variable composed of 4 objects
\itemize{ 
\item norm_matrix: expression or peak matrix (row: gene/peak_location, col: cell) 
\item rowname: gene or peak location 
\item cellname: cell annotation
\item celltype: celltype == sort(unique(cellname))
}
}
\description{
make input format for sctransfer map from matrix file
}
\examples{
data1 <- make_input(input_var) [query case]
data1 <- make_input(input_var, train = TRUE) [training case]

}
