% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{make_input_tm_facs}
\alias{make_input_tm_facs}
\title{Title}
\usage{
make_input_tm_facs(data_list, gene_usage = FALSE, Train = TRUE,
  ncore = 1)
}
\arguments{
\item{data_list}{result of import_tm_facs}
}
\value{

}
\description{
"make_input" for result of import_tm_facs
}
\examples{

}
