% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{make_signature_set}
\alias{make_signature_set}
\title{Title}
\usage{
make_signature_set(signature)
}
\arguments{
\item{signature}{julia object of signature (the result of create_signature)}
}
\value{
initiallized signature set
}
\description{
making signature set
}
\examples{

}
