% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{make_stopwords}
\alias{make_stopwords}
\title{Title}
\usage{
make_stopwords(stopwords)
}
\arguments{
\item{stopwords}{stopword file for word2vec (text file: line by line)}
}
\value{
a julia object of stopword file
}
\description{
make julia object of stopwords for word2vec
}
\examples{

}
