% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{make_w2v_sideinfo}
\alias{make_w2v_sideinfo}
\title{make_w2v_sideinfo}
\usage{
make_w2v_sideinfo(
  w2v_obj,
  cell_ontology_obj,
  stopword,
  norm = "mean",
  R_output = FALSE
)
}
\arguments{
\item{w2v_obj}{the result of make_w2v}

\item{cell_ontology_obj}{the result of make_cellontology}

\item{stopword}{the result of make_stopwords}

\item{norm}{normalization method for word2vector
\itemize{
\item "mean"
\item "median"
\item "max"
\item "weighted_sum": a + b/2 + c/3 + d/4 + ... (for sorted scores)
\item "weighted_abs_sum": a + b/2 + c/3 + d/4 + ... (for absoluted sorted scores)
}}
}
\value{
side information of cell_ontology in julia object
\itemize{
\item field(side_info, "sdatatype"): "w2v"
\item field(side_info, "s"): score matrix
\item field(side_info, "stypenames"): cell types of w2v
\item field(side_info, "snames"): ["NA"]
\item field(side_info, "fjlts"): identity matrix
}
}
\description{
generating a side information of word2vec
}
\examples{
w2v <- make_w2v("Pubmed.index")
cell_ontology <- make_cellontology("cell_ontogloy_file")
stopword <- make_stopwords(stopwords)
side_info <- make_w2v_sideinfo(w2v, cell_ontology, stopword)

}
