% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{n_fold_logistic}
\alias{n_fold_logistic}
\title{Title}
\usage{
n_fold_logistic(tm_facs, Gene_usage, s, annot, mincell, Ngene, data_name,
  sdatatype, fdrthr = 0.1, n = 5, k = 1, tissue_opt = FALSE,
  Celltype_specific = FALSE, ncore = 1)
}
\arguments{
\item{tm_facs}{result of import_tm_facs}
}
\value{
att_set, meta_set, logistic_result
\itemize{
\item att_set: attribute_set
\item meta_set: result of logistic regression (same as create_meta_index_av)
\item logistic_result: result of logistic regression prediction (same a leave_one_val)
\item n_fold: all the results needs n-fold index (ex: res$att_set$1_fold: att_set)
}
}
\description{
Description
}
\examples{
tm_facs <- input_tm_facs("address")
w2v <- make_w2v("Pubmed.index")
cell_ontology <- make_cellontology("cell_ontogloy_file")
stopword <- make_stopwords(stopwords)
side_info <- make_w2v_sideinfo(w2v, cell_ontology, stopword)
res <- n_fold_logistic(tm_facs, 2048, side_info, cell_ontology, FALSE, 64, "user", "cell_ontology")

}
