% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{n_fold_logsitic}
\alias{n_fold_logsitic}
\title{Title}
\usage{
n_fold_logsitic(tm_facs, Gene_usage, s, annot, mincell, Ngene,
  fdrthr = 0.1, n = 5, k = 1, tissue_opt = FALSE,
  Celltype_specific = FALSE, ncore = 1)
}
\arguments{
\item{tm_facs}{result of import_tm_facs}
}
\value{
tm_att_result, tm_meta_result, tm_logistic_result
res[[1]]: tm_att_result -> attribute_set
res[[2]]: tm_meta_result -> result of logistic regression (same as create_meta_index)
res[[3]]: tm_logistic_result -> result of logistic regression prediction (same a leave_one_val)
all the result should decide n-fold index (ex: res[[1]][[1]]: tm_att_result, res[[2]][[5]]: tm_meta_result)
}
\description{
Description
}
\examples{

}
