% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{n_fold_zsl}
\alias{n_fold_zsl}
\title{n_fold_zsl}
\usage{
n_fold_zsl(
  tm_facs,
  Gene_usage,
  data_name,
  side_info,
  w2v_index = TRUE,
  sdatatype = "w2v",
  mincellcount = 10,
  Ngene = 64,
  fdrthr = 0.1,
  simthres = -Inf,
  n = 5,
  side_limit = FALSE,
  norm = "log",
  side_norm = "mean"
)
}
\arguments{
\item{tm_facs}{list variable of importdata (same as import_tm_facs or import_sparse or list of importdata)}

\item{data_name}{name for the data}

\item{side_info}{matrix or w2v object depends on w2v_index
\itemize{
\item (gene or peak) matrix of side information (if w2v_index = FALSE)
\item w2v_side: result of make_w2v_sideinfo (if w2v_index = TRUE)
}}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object of side-information exists}

\item{sdatatype}{type of side-information
\itemize{
\item w2v (cell ontology)
\item scRNA-seq
\item scATAC-seq
}}

\item{mincellcount}{minimum cell count for a given cell type for training}

\item{Ngene}{a number of vector size after JL transformation}

\item{fdrthr}{fdr threshold for calculating similarity score during training}

\item{simthres}{initial similarity threshold}

\item{n}{the number for "N"-fold validation}

\item{side_limit}{TRUE if user wants to limit their side information as cell types in input data, respectively (and min cell fraction threshold)}

\item{norm}{normalization options for matrix
\itemize{
\item "log": log-normalization (with pseudo count)
\item "raw": just uses count matrix
\item "total": divide by total count of each cell, respectively
\item "z": z-normalization
\item "total-log": "total" -> "log" normalization (with pseudo count)
}}

\item{side_norm}{normalization method for word2vector
\itemize{
\item "mean"
\item "median"
\item "max"
\item "weighted_sum": a + b/2 + c/3 + d/4 + ... (for sorted scores)
\item "weighted_abs_sum": a + b/2 + c/3 + d/4 + ... (for absoluted sorted scores)
}}

\item{gene_usage}{option for gene/peak_location filtering to make hadamard matrix}
}
\value{
n-fold list composed of 6 results: conf_list, conf_ratio_list, val_list, tissue_list, f1_list, unassign_list
\itemize{
\item n_fold: all the results needs n-fold index (ex: res$1_fold$conf_list: conf_list)
\item conf_list: confusion matrix
\item conf_ratio_list: ratio_list (confusion_ratio_matrix)
\item val_list: validation matrix
\item tissue_list
\item f1_list: f1 score
\item unassign_list
}
}
\description{
N-fold validation for ZSL.
}
\examples{
1) Side information: Cell Ontology
tm_facs <- input_tm_facs("address")
w2v <- make_w2v("Pubmed.index")
cell_ontology <- make_cellontology("cell_ontogloy_file")
stopword <- make_stopwords(stopwords)
side_info <- make_w2v_sideinfo(w2v, cell_ontology, stopword)
res <- n_fold_zsl(tm_facs, 2048, "user", side_info)

2) Side information: scRNA-seq (or scATAC-seq, etc ...)
tm_facs <- input_tm_facs("address")
sideinfo <- make_input(matrix, train = TRUE)
res <- n_fold_zsl(tm_facs, 2048, "user", side_info, w2v_index = FALSE, sdatatype = "scRNA-seq")

}
