% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{predict_cell}
\alias{predict_cell}
\title{Title}
\usage{
predict_cell(iscell, query, simthr = Inf, gn_overlap_thr = 0.5,
  R_output = TRUE)
}
\arguments{
\item{iscell}{attribute set}

\item{query}{signature object of query (the result of create_signature)}

\item{simthr}{default simthreshold for prediction (default: Inf -> considers only simthr in attribute set)}

\item{gn_overlap_thr}{minimum gene overlap ratio of attribute in attribute set between query data (default: 0.5)}

\item{R_output}{TRUE ifthe  result of the function should be R object. FALSE if the result of the function should be julia object (default: TRUE)}
}
\value{
a confusion matrix of prediction result
row: query cell type or cluster name
column: cell types in side-information (from query)
first column: unassign count
}
\description{
make a confusion matrix of query data by ZSL with consensus manner
}
\examples{

}
