% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{quick_prediction}
\alias{quick_prediction}
\title{quick_prediction}
\usage{
quick_prediction(input_var, name, opt = "consensus")
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column)\cr
- same as result of importdata or convert_seurat}

\item{name}{any name that user wants for the result}

\item{opt}{prediction mode whether "consensus" or "logistic"}
}
\value{
a result of predicted cell types for each query cell cluster
}
\description{
user-friendly quick prediction for single-cell data
side-information is cell ontology
}
\examples{
result <- quick_prediction(matrix, "single-cell")
- prediction by consensus results of ZSL
result <- quick_prediction(matirx, "single-cell", opt = "logistic")
- prediction by logistic regression

}
