% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{quick_training}
\alias{quick_training}
\title{Title}
\usage{
quick_training(input_var, name, tissue)
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column) (same as result of importdata or convert_seurat)}

\item{name}{any name that user wants for the result}

\item{tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter) 
Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
quick traing of zero-shot-learing (ZSL)
}
\examples{

}
