% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{r2julia_sort}
\alias{r2julia_sort}
\title{r2julia_sort}
\usage{
r2julia_sort(r_list, julia_list, Index = TRUE)
}
\arguments{
\item{r_list}{list that want to change}

\item{julia_list}{target list}

\item{Index}{TRUE if the result should be an index for r_list. FALSE if the result should be the changed list or r_list}
}
\value{
index or changed list of r_list
}
\description{
make a index for a given list order
}
\examples{
r_list
new_list
new_index <- r2julia_sort(r_list, new_list)
r_list_new <- r_list[new_index]

}
