% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{save_sparse_data}
\alias{save_sparse_data}
\title{save_sparse_data}
\usage{
save_sparse_data(input_var, filename)
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column)\cr
- same as result of importdata or convert_seurat}

\item{filename}{filename to be saved}
}
\value{
no return value. But, it saves sparse matrix in MatrixMarket format with cell annotation file (for column name) and gene name or peak location, etc... (for rowname)
}
\description{
save MatrixMarket format of sparse matrix
}
\examples{
data <- importdata("user_file")
save_sparse_data(data, "user_scData")

}
