% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{sparse_tm_facs}
\alias{sparse_tm_facs}
\title{sparse_tm_facs}
\usage{
sparse_tm_facs(tm_facs)
}
\arguments{
\item{tm_facs}{list variable of importdata (same as import_tm_facs or import_sparse or list of importdata)}
}
\value{
save files in sparse format (see save_sparse_data)
}
\description{
"save_sparse_data" for multiple "raw" matrix object
}
\examples{
tm_facs <- import_tm_facs("address_file")
sparse_tm_facs(tm_facs)

}
