% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{tm_facs_logistic}
\alias{tm_facs_logistic}
\title{Title}
\usage{
tm_facs_logistic(att_set, meta_model, sig_set, Tissue_opt = FALSE,
  celltype_specific = FALSE)
}
\arguments{
\item{description}{of parameters}
}
\value{

}
\description{
Description
}
\examples{

}
