% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{validation}
\alias{validation}
\title{validation}
\usage{
validation(conf, opt = FALSE)
}
\arguments{
\item{conf}{confusion ratio matrix from confusion_ratio or just confusion matrix from predict_cell}

\item{opt}{TRUE if conf is a confusion matrix and FALSE if conf is a confusion_ratio matrix}
}
\value{
True postivie rate and unassign ratio for each cell type, respectively
}
\description{
validation of known data
}
\examples{
confusion <- predict_cell(att_set, signature)
conf_val <- validation(confusion, opt = TRUE)

}
