% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{validation_sample}
\alias{validation_sample}
\title{validation_sample}
\usage{
validation_sample(conf)
}
\arguments{
\item{conf}{confusion matrix from predict_cell}
}
\value{
f1_score, unassign
\itemize{
\item f1_score: list of F1 scores for each sample
\item unassign: list of unassign ratio for each sample
}
}
\description{
calculate F1 scores and unassign ratio
}
\examples{
conf <- predict_cell(att_set, query_signature)
res <- validation_sample(conf)

}
