% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{add_attribute_set}
\alias{add_attribute_set}
\title{add_attribute_set}
\usage{
add_attribute_set(attribute, iscell)
}
\arguments{
\item{attribute}{julia object of attribute (the result of create_attribute or quick_training)}

\item{iscell}{pre-existing attribute set (the result of make_attribute_set)}
}
\value{
merged attribute set
}
\description{
adding attribute object into attribute set
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
att <- create_attribute("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)
att_set <- add_attribute_set(att, att_set)
- att_set: pre-existing attribute-set

}
