% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{add_signature_set}
\alias{add_signature_set}
\title{add_signature_set}
\usage{
add_signature_set(signature, mscell)
}
\arguments{
\item{signature}{julia object of signature (the result of create_signature)}

\item{mscell}{per-existing signature set (the result of make_signature_set)}
}
\value{
merged signature set
}
\description{
adding signature object into signature set
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
signature <- create_signature("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)
sig_set <- add_signature_set(signature, sig_set)
- sig_set: pre-existing signature set

}
