% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{confusion_ratio}
\alias{confusion_ratio}
\title{Title}
\usage{
confusion_ratio(confusion)
}
\arguments{
\item{confusion}{confusion matrix from predict_cell}
}
\value{
ratio of each query cell type
}
\description{
convert count confusion matrix into ratio of each query cell type, respectively
}
\examples{

}
