% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{create_attribute}
\alias{create_attribute}
\title{create_attribute}
\usage{
create_attribute(Name, Tissue, w2v, Expr, Rowname, Cellname, cell_ontology,
  Sdatatype, w2v_index = FALSE, Xdatatype = "scRNA-seq", Ngene = 64,
  Stopwords = stopwords, Fdrthr = 0.1, Gamma = 1, Lambda = 1,
  Xpeaks = FALSE, Sep = "_", Mincellcount = 10, Mincellfrac = 0.05,
  Simthres = -Inf, R_output = FALSE)
}
\arguments{
\item{Name}{user-defined name for the output}

\item{Tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
\itemize{
\item Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea
}}

\item{w2v}{word2vector object
\itemize{
\item output of make_w2v
\item pre-existing side-information object (julia object) (ex: julia_obj$side_information) 
\item output of generate_sideinformation
}}

\item{Expr}{expression/peak matrix (d <- make_input(input_var), d$norm_matrix)}

\item{Rowname}{gene/peak_location of Expr (d<- make_input(input_var), d$rowname)}

\item{Cellname}{cellname of Expr (d<- make_input(input_var), d$cellname)}

\item{cell_ontology}{julia object of cell ontology (the result of make_cellontology)}

\item{Sdatatype}{type of side-information (cell_ontology, ATAC-seq, or user-defined term)}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object of side-information exists}

\item{Xdatatype}{type of Expr (scRNA-seq or scATAC-seq)}

\item{Ngene}{a number of vector size after JL transformation}

\item{Stopwords}{julia object of stopwords (the result of make_stopwords)}

\item{Fdrthr}{fdr threshold for calculating similarity score during training}

\item{Gamma}{paramter for ZSL}

\item{Lambda}{parameter for ZSL}

\item{Xpeaks}{TRUE if data type of x (and Expr) is ATAC-seq}

\item{Sep}{delimiter for peak_location\cr
ex: chr1_3084739_3085712 (make sure delimiter for peak_location of data must be "_")}

\item{Mincellcount}{minimum cell count for a given cell type for training}

\item{Mincellfrac}{minimum fraction of cell type for training}

\item{Simthres}{initial similarity threshold}

\item{R_output}{TRUE if the result of the function should be R object. FALSE if the result of the function should be julia object}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
make attribute object for ZSL (trained data)
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
att <- create_attribute("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)

}
