% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{create_attribute}
\alias{create_attribute}
\title{create_attribute}
\usage{
create_attribute(
  Name,
  Tissue,
  Expr,
  Rowname,
  Cellname,
  side_info,
  w2v_index = TRUE,
  snames = c(),
  stypenames = c(),
  sdim = 200,
  Xdatatype = "scRNA-seq",
  Ngene = 64,
  Fdrthr = 0.1,
  Gamma = 1,
  Lambda = 1,
  Sep = "_",
  Sdatatype = "w2v",
  Mincellcount = 10,
  Mincellfrac = 0,
  Simthres = -Inf,
  norm = "mean",
  R_output = FALSE
)
}
\arguments{
\item{Name}{user-defined name for the output}

\item{Tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
\itemize{
\item Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea
}}

\item{Expr}{expression/peak matrix (d <- make_input(input_var), d$norm_matrix)}

\item{Rowname}{gene/peak_location of Expr (d<- make_input(input_var), d$rowname)}

\item{Cellname}{cellname of Expr (d<- make_input(input_var), d$cellname)}

\item{side_info}{matrix or w2v object depends on w2v_index
\itemize{
\item (gene or peak) matrix of side information (if w2v_index = FALSE)
\item w2v_side: result of make_w2v_sideinfo (if w2v_index = TRUE)
}}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object of side-information exists}

\item{snames}{gene/peak_location of matrix for side information (if w2v_index = FALSE)}

\item{stypenames}{cellname of matrix for side information (if w2v_index = FALSE)}

\item{sdim}{a number of vector size after JL transformation for side information}

\item{Xdatatype}{type of Expr 
\itemize{
\item scRNA-seq 
\item scATAC-seq
}}

\item{Ngene}{a number of vector size after JL transformation for "X" matrix}

\item{Fdrthr}{fdr threshold for calculating similarity score during training}

\item{Gamma}{paramter for ZSL}

\item{Lambda}{parameter for ZSL}

\item{Sep}{delimiter for peak_location\cr
ex: chr1_3084739_3085712 (make sure delimiter for peak_location of data must be "_")}

\item{Sdatatype}{type of side-information 
\itemize{
\item w2v (cell ontology)
\item scRNA-seq
\item scATAC-seq
}}

\item{Mincellcount}{minimum cell count for a given cell type for training}

\item{Mincellfrac}{minimum fraction of cell type for training}

\item{Simthres}{initial similarity threshold}

\item{norm}{normalization method for word2vector
\itemize{
\item "mean"
\item "median"
\item "max"
\item "weighted_sum": a + b/2 + c/3 + d/4 + ... (for sorted scores)
\item "weighted_abs_sum": a + b/2 + c/3 + d/4 + ... (for absoluted sorted scores)
}}

\item{R_output}{TRUE if the result of the function should be R object. FALSE if the result of the function should be julia object}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
make attribute object for ZSL (trained data)
}
\examples{
1) Side information: Cell Ontology
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
w2v_side <- make_w2v_sideinfo(w2v, annot, stopwords)
att <- create_attribute("user", "Heart", d$norm_matirx, d$rowname, d$cellname, w2v_side)

2) Side information: scRNA-seq (or scATAC-seq, etc...)
d <- make_input(matrix1, train = TRUE)
sideinfo <- make_input(matrix2, train = TRUE)
att <- create_attribute("user", "Heart", d$norm_matirx, d$rowname, d$cellname, sideinfo$norm_matrix, w2v_index = FALSE, snames = sideinfo$rowname, stypenames = sideinfo$cellname, sdatatype = "scRNA-seq")

}
