% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{create_signature}
\alias{create_signature}
\title{Title}
\usage{
create_signature(Name, Tissue, w2v, Expr, Genename, Cellname,
  cell_ontology, Sdatatype, w2v_index = FALSE, Stopwords = stopwords,
  Xdatatype = "genes", R_output = FALSE, celltype_cnt_thr = TRUE)
}
\arguments{
\item{Name}{user-defined name for the output}

\item{Tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea}

\item{w2v}{word2vector object
output of make_w2v
pre-existing side-information object (julia object) (ex: julia_obj$side_information)
output of generate_sideinformation}

\item{Expr}{expression/peak matrix (first index from the result of make_input: d <- make_input(input_var), d[[1]])}

\item{Genename}{gene/peak_location of Expr (second index from the result of make_input: d<- make_input(input_var), d[[2]])}

\item{Cellname}{cellname of Expr (third index from the result of make_input: d<- make_input(input_var), d[[3]])}

\item{cell_ontology}{julia object of cell ontology (the result of make_cellontology)}

\item{Sdatatype}{type of side-information (cell_ontology, ATAC-seq, or user-defined term)}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object for side-information exists (default: FALSE)}

\item{Stopwords}{julia object of stopwords (the result of make_stopwords) (default: pre-existing variable: stopwords)}

\item{Xdatatype}{type of Expr (default: scRNA-seq; or scATAC-seq)}

\item{R_output}{TRUE ifthe  result of the function should be R object. FALSE if the result of the function should be julia object (default: FALSE)}

\item{celltype_cnt_thr}{if the number for cell type exceeds, warning message will be sent and return FALSE. To ignore it, change the parameter into FALSE}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
make signature object for ZSL (input for prediction)
}
\examples{

}
