% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{create_signature}
\alias{create_signature}
\title{create_signature}
\usage{
create_signature(Name, Tissue, w2v, Expr, Rowname, Cellname, cell_ontology,
  Sdatatype, w2v_index = FALSE, Stopwords = stopwords,
  Xdatatype = "scRNA-seq", R_output = FALSE, celltype_cnt_thr = TRUE,
  mincellfrac = 0.01, mincellcount = 10, side_limit = FALSE)
}
\arguments{
\item{Name}{user-defined name for the output}

\item{Tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
\itemize{
\item Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea
}}

\item{w2v}{word2vector object
\itemize{
\item output of make_w2v
\item pre-existing side-information object (julia object) (ex: julia_obj$side_information)
\item output of generate_sideinformation
}}

\item{Expr}{expression/peak matrix (d <- make_input(input_var), d$norm_matrix)}

\item{Rowname}{gene/peak_location of Expr (d<- make_input(input_var), d$rowname)}

\item{Cellname}{cellname of Expr (d<- make_input(input_var), d$cellname)}

\item{cell_ontology}{julia object of cell ontology (the result of make_cellontology)}

\item{Sdatatype}{type of side-information (cell_ontology, ATAC-seq, or user-defined term)}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object of side-information exists}

\item{Stopwords}{julia object of stopwords (the result of make_stopwords)}

\item{Xdatatype}{type of Expr (scRNA-seq or scATAC-seq)}

\item{R_output}{TRUE if the result of the function should be R object. FALSE if the result of the function should be julia object}

\item{celltype_cnt_thr}{if the number for cell type exceeds, warning message will be sent and return FALSE. To ignore it, change the parameter into FALSE}

\item{Mincellcount}{minimum cell count for a given cell type (works when side_limit: TRUE)}

\item{Mincellfrac}{minimum fraction of cell type (works when side_limit: TRUE)}

\item{side_liimit}{TRUE if user wants to limit their side information as cell types in input data. Otherwise, it will use all the cell types in side information (= FALSE)}
}
\value{
an attribute for ZSL (trained data) (julia object)
}
\description{
make signature object for ZSL (input for prediction)
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
signature <- create_signature("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)

}
