% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{importdata}
\alias{importdata}
\title{Description of importdata}
\usage{
importdata(file, opt = "raw", datatype = "RNAseq",
  colname_file = FALSE, rowname_file = FALSE)
}
\arguments{
\item{file}{file of the single-cell data (ex: scRNA-seq, scATAC)}

\item{opt}{data format of the single-cell data ("raw": tab-delimited text file, "SingleCellExperiment": SingleCellExperiment rds file, "sparse": MatrixMarket format (written by writeMM)}

\item{datatype}{"RNAseq" or "atac"}

\item{colname_file}{colname_file for sparse matrix (usually cell name), when opt is "sparse". Otherwise, it is FALSE}

\item{rowname_file}{rowname_file for sparse matrix (gene name: official gene symbol or peak location), when opt is "sparse". Otherwise it is FALSE}
}
\value{
a matrix class composed of gene/peak_location(row)-by-cell(column)
}
\description{
import_data will import 3 types of single-cell data format to create gene-by-cell matrix
}
\examples{
data <- importdata("filename") (default: opt = "raw", datatype = "RNAseq", colname_file = FALSE, rowname_file = FALSE)

}
