% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{julia_environment}
\alias{julia_environment}
\title{julia_environment}
\usage{
julia_environment(julia_obj_index = FALSE)
}
\arguments{
\item{julia_obj_index}{it should be TRUE if user wants to import pre-existing julia object}
}
\value{
return the julia_object if julia_obj_index is TRUE. Otherwise, there is no output \cr
- julia_object
\itemize{
\item side_information: word2vector
\item cell_ontology: description of cell types
\item attribute_set: 'V' matrix
\item signature_set: for logistic regression modeling
\item meta_model: logistic regression model
}
}
\description{
setup the environment for calling Julia packages and Julia code (CoreFunction.jl, importing pre-existing julia object is optional)
}
\examples{
julia_environment()
julia_obj <- julia_environment(julia_obj_index = TRUE)

}
