% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{leave_one_val}
\alias{leave_one_val}
\title{leave_one_val}
\usage{
leave_one_val(att_set, sig_set)
}
\arguments{
\item{att_set}{attribute set}

\item{sig_set}{signature set}
}
\value{
6 results: conf_list, conf_ratio_list, val_list, tissue_list, f1_list, unassign_list
\itemize{
\item conf_list: confusion matrix
\item conf_ratio_list: ratio_list (confusion_ratio_matrix)
\item val_list: validation matrix
\item tissue_list
\item f1_list: f1 score
\item unassign_list
}
- each result is composed of sample in TM (ex: res$conf_list[1]: confusion matrix of Aorta tissue)
}
\description{
leave one tissue validation of TM (the parameter for side-information of att_set and sig_set must be same)
}
\examples{
res <- leave_one_val(att_set, sig_set)

}
