% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{logistic_prediction_cell}
\alias{logistic_prediction_cell}
\title{logistic_prediciton_cell}
\usage{
logistic_prediction_cell(iscell, meta_model_list, query,
  gn_overlap_thr = 0.5, logistic_thr = 0.5, tissue_opt = FALSE,
  celltype_specific = FALSE)
}
\arguments{
\item{iscell}{attribute set}

\item{meta_model_list}{result from create_meta_index}

\item{query}{signature object of query}

\item{gn_overlap_thr}{minimum gene overlap ratio of attribute in attribute set between query data}

\item{logistic_thr}{threshold for each cell whether predicted score from logistic regression model is confident}

\item{tissue_opt}{TRUE if user wants to find tissue-origin}

\item{celltype_specific}{TRUE if user wants to calculate the weight fo each tissue with only matched cell type. Otherwise, weight for each tissue will be calculated with whole celltypes (default: FALSE; tissue_opt should be TRUE to use this parameter)}
}
\value{
return 2 confusion matrix 
\itemize{
\item confusion: a confusion matrix of prediction result
\item all_confusion: a summarized matrix of prediction result in each cell
}
}
\description{
predict cell type with logistic regression model
}
\examples{
meta_model <- create_meta_index_av(att_set, sig_set)
res_list <- logistic_prediction_cell(att_set, meta_model, query_signature)
res_list$confusion or res_list$all_confusion

}
