% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{make_attribute_set}
\alias{make_attribute_set}
\title{Title}
\usage{
make_attribute_set(attribute)
}
\arguments{
\item{attribute}{julia object of attribute (the result of create_attribute or quick_training)}
}
\value{
initiallized attribute set
}
\description{
making attribute set
}
\examples{

}
