% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{make_cellontology}
\alias{make_cellontology}
\title{make_cellontology}
\usage{
make_cellontology(cell_ontology, celltype = c())
}
\arguments{
\item{cell_ontology}{cell_ontology file which is composed of celltype and description of corresponding celltype separated by |}

\item{celltype}{extract the cell_ontology with a given celltype list or use all the celltypes 
\itemize{
\item c(): use all the cell types in cell_ontology file
\item celltype -> celltype: use only the cell types in cellname variable
}}
}
\value{
a julia object of cell ontology
}
\description{
make cell ontology variable from cell ontology file
}
\examples{
cell_ontology <- make_cellontology("cell_ontology_file")
- using all the cell types in cell_ontology file

d <- make_input(input_var)
cell_ontology <- make_cellontology("cell_ontology_file", celltype = d$celltype)
- using only the cell types in the "d"

}
