% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{make_input}
\alias{make_input}
\title{Title}
\usage{
make_input(input_var, min_thr = 0.1, max_thr = 0.9, pseudo_cnt = 1,
  train = FALSE, gene_usage = FALSE)
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column) (same as result of importdata or convert_seurat)}

\item{min_thr}{minimum threshold of expressing cell ratio for a given gene/peak_location (default: 0.1)}

\item{max_thr}{maximum threshold of expressing cell ratio for a given gene/peak_location (default: 0.9)}

\item{pseudo_cnt}{pseudo count for count matrix for log-transformation (default: 1)}

\item{train}{TRUE if it is data for training (default: FALSE) 
train -> TRUE: adjust min_thr and max_thr. In addition, make the gene/peak_location count for hadamard matrix
train -> FALSE: no gene/peak_location filtering}

\item{gene_usage}{option for gene/peak_location filtering to make hadamard matrix (default: FALSE; train parameter must be TRUE)
gene_usage -> FALSE: gene/peak_location filtering by discarding low expressing cell ratio genes/peak_locations}
}
\value{
return a list variable composed of 4 objects (d, genename, cellname, celltype)
d: expression or peak matrix (row: gene, col: cell)
genename: gene or peak location
cellname: cell annotation
celltype: celltype == sort(unique(cellname))
}
\description{
make input format for sctransfer map from matrix file
}
\examples{
data1 <- make_input(input_var) [query case]
data1 <- make_input(input_var, train = TRUE) [training case]
usage: d -> data1[[1]], genename -> data1[[2]], cellname -> data1[[3]], celltype -> data1[[4]]

}
