% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{make_signature_set}
\alias{make_signature_set}
\title{make_signature_set}
\usage{
make_signature_set(signature)
}
\arguments{
\item{signature}{julia object of signature (the result of create_signature)}
}
\value{
initiallized signature set
}
\description{
making signature set
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
signature <- create_signature("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)
sig_set <- make_signature_set(signature)

}
