% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{make_tm_facs_attribute_set}
\alias{make_tm_facs_attribute_set}
\title{Title}
\usage{
make_tm_facs_attribute_set(tm_facs_input, s, annot, mincell, ngene = 64,
  fdrthr = 0.1)
}
\arguments{
\item{tm_facs_input}{the result of make_input_tm_facs}

\item{s}{side-information (jula object)}

\item{annot}{cell ontology (julia object)}

\item{mincell}{TRUE if user wants to give minimum cell type thresholding}
}
\value{
attribute set
}
\description{
creating attribute set from the result of make_input_tm_facs
}
\examples{

}
