% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{make_tm_facs_signature_set}
\alias{make_tm_facs_signature_set}
\title{Title}
\usage{
make_tm_facs_signature_set(tm_facs_input, s, annot)
}
\arguments{
\item{tm_facs_input}{result of make_input_tm_facs}

\item{s}{side-infomration (julia object)}

\item{annot}{cell_ontology (julia object)}
}
\value{
signature set
}
\description{
creating signature set from result of make_input_tm_facs
}
\examples{

}
