% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{n_fold_logistic}
\alias{n_fold_logistic}
\title{n_fold_logistic}
\usage{
n_fold_logistic(
  tm_facs,
  Gene_usage,
  data_name,
  side_info,
  w2v_index = TRUE,
  sdatatype = "w2v",
  mincellcount = 10,
  Ngene = 64,
  fdrthr = 0.1,
  simthres = -Inf,
  n = 5,
  k = 1,
  tissue_opt = FALSE,
  Celltype_specific = FALSE,
  ncore = 1,
  side_limit = FALSE,
  norm = "log",
  side_norm = "mean"
)
}
\arguments{
\item{tm_facs}{list variable of importdata (same as import_tm_facs or import_sparse or list of importdata)}

\item{data_name}{name for the data}

\item{side_info}{matrix or w2v object depends on w2v_index
\itemize{
\item (gene or peak) matrix of side information (if w2v_index = FALSE)
\item w2v_side: result of make_w2v_sideinfo (if w2v_index = TRUE)
}}

\item{w2v_index}{TRUE if user needs to make side-information from cell ontology. FALSE if julia object of side-information exists}

\item{sdatatype}{type of side-information
\itemize{
\item w2v (cell ontology)
\item scRNA-seq
\item scATAC-seq
}}

\item{mincellcount}{minimum cell count for a given cell type for training}

\item{Ngene}{a number of vector size after JL transformation}

\item{fdrthr}{fdr threshold for calculating similarity score during training}

\item{n}{the number for "N"-fold validation}

\item{k}{ratio between ZSL(1) and logistic regression modeling (k) (k=1 -> ZSL:glm = 1:1, k=2 -> ZSL:glm = 1:2)}

\item{tissue_opt}{TRUE if user wants to find tissue-origin}

\item{Celltype_specific}{TRUE if user wants to calculate the weight fo each tissue with only matched cell type. Otherwise, weight for each tissue will be calculated with whole celltypes (deault: FALSE; tissue_opt should be TRUE to use this parameter)}

\item{ncore}{the number of cores for multi-threading}

\item{side_limit}{TRUE if user wants to limit their side information as cell types in input data, respectively (and min cell fraction threshold)}

\item{norm}{normalization options for matrix
\itemize{
\item "log": log-normalization (with pseudo count)
\item "raw": just uses count matrix
\item "total": divide by total count of each cell, respectively
\item "z": z-normalization
\item "total-log": "total" -> "log" normalization (with pseudo count)
}}

\item{side_norm}{normalization method for word2vector
\itemize{
\item "mean"
\item "median"
\item "max"
\item "weighted_sum": a + b/2 + c/3 + d/4 + ... (for sorted scores)
\item "weighted_abs_sum": a + b/2 + c/3 + d/4 + ... (for absoluted sorted scores)
}}

\item{gene_usage}{option for gene/peak_location filtering to make hadamard matrix}
}
\value{
att_set, meta_set, logistic_result
\itemize{
\item att_set: attribute_set
\item meta_set: result of logistic regression (same as create_meta_index_av)
\item logistic_result: result of logistic regression prediction (same a leave_one_val)
\item n_fold: all the results needs n-fold index (ex: res$att_set$1_fold: att_set)
}
}
\description{
N-fold validation for logistic regression model. Uses training data for ZSL and logistic regression modeling independently (determined by "k").
}
\examples{
1) Side information: Cell Ontology
tm_facs <- input_tm_facs("address")
w2v <- make_w2v("Pubmed.index")
cell_ontology <- make_cellontology("cell_ontogloy_file")
stopword <- make_stopwords(stopwords)
side_info <- make_w2v_sideinfo(w2v, cell_ontology, stopword)
res <- n_fold_logistic(tm_facs, 2048, "user", side_info)

2) Side information: scRNA-seq (or scATAC-seq, etc ...)
sideinfo <- make_input(matrix, train = TRUE)
res <- n_fold_logistic(tm_facs, 2048, "user", side_info, w2v_index = FALSE, sdatatype = "scRNA-seq")

}
