% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_code.R
\name{n_fold_zsl}
\alias{n_fold_zsl}
\title{n_fold_zsl}
\usage{
n_fold_zsl(tm_facs, Gene_usage, s, annot, mincell, Ngene, data_name,
  sdatatype, fdrthr = 0.1, n = 5, side_limit = FALSE, norm = "log")
}
\arguments{
\item{tm_facs}{list variable of importdata (same as import_tm_facs or import_sparse or list of importdata)}

\item{s}{side-infomration (julia object)}

\item{annot}{cell_ontology (julia object)}

\item{mincell}{TRUE if user wants to give minimum cell type thresholding}

\item{Ngene}{a number of vector size after JL transformation}

\item{data_name}{name for the attribute}

\item{sdatatype}{type of side-information (cell_ontology, ATAC-seq, or user-defined term)}

\item{fdrthr}{fdr threshold for calculating similarity score during training}

\item{n}{the number for "N"-fold validation}

\item{side_limit}{TRUE if user wants to limit their side information as cell types in input data, respectively (and min cell fraction threshold)}

\item{norm}{normalization options for matrix
\itemize{
\item "log": log-normalization (with pseudo count)
\item "raw": just uses count matrix
\item "total": divide by total count of each cell, respectively
\item "z": z-normalization
\item "total-log": "total" -> "log" normalization (with pseudo count)
}}

\item{gene_usage}{option for gene/peak_location filtering to make hadamard matrix}
}
\value{
n-fold list composed of 6 results: conf_list, conf_ratio_list, val_list, tissue_list, f1_list, unassign_list
\itemize{
\item n_fold: all the results needs n-fold index (ex: res$1_fold$conf_list: conf_list)
\item conf_list: confusion matrix
\item conf_ratio_list: ratio_list (confusion_ratio_matrix)
\item val_list: validation matrix
\item tissue_list
\item f1_list: f1 score
\item unassign_list
}
}
\description{
N-fold validation for ZSL.
}
\examples{
tm_facs <- input_tm_facs("address")
w2v <- make_w2v("Pubmed.index")
cell_ontology <- make_cellontology("cell_ontogloy_file")
stopword <- make_stopwords(stopwords)
side_info <- make_w2v_sideinfo(w2v, cell_ontology, stopword)
res <- n_fold_zsl(tm_facs, 2048, side_info, cell_ontology, FALSE, 64, "user", "cell_ontology")

}
