% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{predict_cell}
\alias{predict_cell}
\title{predict_cell}
\usage{
predict_cell(
  iscell,
  query,
  simthr = Inf,
  gn_overlap_thr = 0.5,
  R_output = TRUE
)
}
\arguments{
\item{iscell}{attribute set}

\item{query}{signature object of query (the result of create_signature)}

\item{simthr}{default simthreshold for prediction (default: Inf -> considers only simthr in attribute set)}

\item{gn_overlap_thr}{minimum gene overlap ratio of attribute in attribute set between query data}

\item{R_output}{TRUE if the result of the function should be R object. FALSE if the result of the function should be julia object}
}
\value{
return 2 confusion matrix
\itemize{
\item confusion: a confusion matrix of prediction result
\item all_confusion: a summarized matrix of prediction result in each cell
}
}
\description{
make a confusion matrix of query data by ZSL with consensus manner
}
\examples{
d <- make_input(matrix, train = TRUE)
w2v <- make_w2v("Pubmed_index")
annot <- make_cellontology("cell ontology file")
stopwords <- make_stopwords(stopwords)
signature <- create_signature("user", "Heart", d$norm_matirx, d$rowname, d$cellname, annot, "cell_ontology", w2v_index = TRUE)
confusion <- predict_cell(att_set, signature)
- att_set: pre-existing attribute set

}
