% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{quick_prediction}
\alias{quick_prediction}
\title{quick_prediction}
\usage{
quick_prediction(input_var, name, tissue, xdatatype, opt = "consensus")
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column)\cr
- same as result of importdata or convert_seurat}

\item{name}{any name that user wants for the result}

\item{tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is a new tissue type, it doesn't matter)
\itemize{
\item Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea
}}

\item{xdatatype}{type of Expr
\itemize{
\item scRNA-seq
\item scATAC-seq
}}

\item{opt}{prediction mode whether "consensus" or "logistic"}
}
\value{
return 2 confusion matrix
\itemize{
\item confusion: a confusion matrix of prediction result
\item all_confusion: a summarized matrix of prediction result in each cell
}
}
\description{
user-friendly quick prediction for single-cell data
side-information is cell ontology (w2v)
}
\examples{
result <- quick_prediction(matrix, "single-cell", "Heart", "scRNA-seq")
- prediction by consensus results of ZSL
result <- quick_prediction(matirx, "single-cell", "Heart", "scRNA-seq", opt = "logistic")
- prediction by logistic regression

}
