% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{quick_prediction}
\alias{quick_prediction}
\title{Title}
\usage{
quick_prediction(input_var, name, opt = "consensus")
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column) (same as result of importdata or convert_seurat)}

\item{name}{any name that user wants for the result}

\item{opt}{prediction mode whethere "consensus" or "logistic" (default: "consensus")}
}
\value{
a result of predicted cell types for each query cell cluster
}
\description{
user-friendly quick prediction for single-cell data
side-information is cell ontology
}
\examples{

}
