% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{quick_training}
\alias{quick_training}
\title{quick_training}
\usage{
quick_training(input_var, name, tissue)
}
\arguments{
\item{input_var}{matrix class composed of gene/peak_location(row)-by-cell(column)\cr
- same as result of importdata or convert_seurat}

\item{name}{any name that user wants for the result}

\item{tissue}{tissue of the sample (please make sure the word of tissue must be same as pre-existing tissue list. If it is new tissue type, it doesn't matter)
\itemize{ 
\item Mouse: Aorta, Brain, Diaphragm, Fat, Heart, Kidney, LargeIntestine, LimbMuscle, Liver, Lung, MammaryGland, Marrow, Pancreas, Skin, Spleen, Thymus, Tongue, Trachea
}}
}
\value{
an attribute for ZSL (trained data in julia object)
}
\description{
quick traing of zero-shot-learing (ZSL)
}
\examples{
att_heart <- quick_training(matrix, "single-cell", "Heart")
att_lung <- quick_training(matrix, "single-cell", "Lung")
att_set <- make_attribute_set(att_heart)
att_set <- add_attribute_set(att_lung, att_set)

}
