% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods.R
\name{validation}
\alias{validation}
\title{validation}
\usage{
validation(conf_ratio)
}
\arguments{
\item{conf_ratio}{confusion ratio matrix from confusion_ratio}
}
\value{
True postivie rate and unassign ratio for each cell type, respectively
}
\description{
validation of known data
}
\examples{
confusion <- predict_cell(att_set, signature)
conf_ratio <- confusion_ratio(conf_ratio)
conf_val <- validation(conf_ratio)

}
