% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreMethods_av.R
\name{validation}
\alias{validation}
\title{Title}
\usage{
validation(conf_ratio)
}
\arguments{
\item{con_ratio}{confusion ratio matrix from confusion_ratio}
}
\value{
True postivie rate and unassign ratio for each cell type, respectively
}
\description{
validation of known data
}
\examples{

}
